fix bug

+++ iq/day4/arrayQuestions/arrayQuestions.hpp
@@ -7,7 +7,7 @@ namespace iq {
 
 void SetImportantValues(int* a_someImportant, size_t a_firstSize, int* a_allImportant, size_t a_secondSize);
 
-void SetImportantValuesSorted(int* a_someImportant, size_t a_firstSize, int* a_allImportant, size_t a_secondSize);
+void SetImportantValuesSorted(int* a_someImportant, int a_firstSize, int* a_allImportant, int a_secondSize);
 
 } // iq
